move tileset camera and bounding volumes conversion into a different files
